Added authorization policies

+++ app/Http/Controllers/CategoryController.php
@@ -7,6 +7,18 @@ use App\Category;
 use Session;
 class CategoryController extends Controller
 {
+
+    /**
+     * Create a new controller instance.
+     *
+     * @return void
+     */
+    public function __construct()
+    {
+        $this->middleware('auth:admin');
+    }
+
+    
     /**
      * Display a listing of the resource.
      *
@@ -14,6 +26,7 @@ class CategoryController extends Controller
      */
     public function index()
     {
+        $this->authorize('view',new Category());
        $categories = Category::all();
        return view('admin.categories.index')->withCategories($categories);
     }
@@ -25,6 +38,7 @@ class CategoryController extends Controller
      */
     public function create()
     {
+         $this->authorize('create',Category::class);
         $categories = Category::all();
         return view('admin.categories.create')->withCategories($categories);
     }
@@ -38,6 +52,7 @@ class CategoryController extends Controller
      */
     public function store(Request $request)
     {
+         $this->authorize('create',Category::class);
          $this->validate($request,[
             'category'=>'required|max:255|unique:categories,category',
             'description'=>'required'
@@ -63,6 +78,7 @@ class CategoryController extends Controller
     public function show($id)
     {
         $category = Category::find($id);
+         $this->authorize('view',$category);
         return view('admin.categories.show')->withCategory($category);
     }
 
@@ -75,6 +91,7 @@ class CategoryController extends Controller
     public function edit($id)
     {
         $category= Category::find($id);
+         $this->authorize('update',$category);
         return view('admin.categories.edit')->withCategory($category);
     }
 
@@ -93,6 +110,8 @@ class CategoryController extends Controller
             ]);
 
         $category = Category::find($id);
+
+         $this->authorize('update',$category);
         $category->category=$request->category;
         $category->description=$request->description;
         $category->save();
@@ -111,7 +130,8 @@ class CategoryController extends Controller
      */
     public function destroy($id)
     {
-        $category = Category::find($id); 
+        $category = Category::find($id);
+         $this->authorize('delete',$category); 
         $category->delete();
         Session::flash('success','The category was successfully deleted !');
         return redirect()->route('categories.index');
